requirements.txt
streamlit
tensorflow-cpu
keras
numpy
pandas
requests
googledrivedownloader
nltk
altair
seaborn
plotly